+++ packages/gatsby/src/commands/develop-process.ts
@@ -101,15 +101,6 @@ module.exports = async (program: IDevelopArgs): Promise<void> => {
     }
   )
 
-  if (process.env.GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES) {
-    reporter.panic(
-      `The flag ${chalk.yellow(
-        `GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES`
-      )} is not available with ${chalk.cyan(
-        `gatsby develop`
-      )}, please retry using ${chalk.cyan(`gatsby build`)}`
-    )
-  }
   initTracer(program.openTracingConfigFile)
   markWebpackStatusAsPending()
   reporter.pendingActivity({ id: `webpack-develop` })